Add custom unmarshaller for Module.Use to improve error message
* New error message:
```
failed to parse the blueprint in tst.yaml, check YAML syntax for errors, err=line 32: `use` must be a list of module ids
```
* Move YAML-related code `pkg/config` into `yaml*.go`.

NOTE to reviewer:
Most of the changed lines a just moved.
The new code is in following functions:

* `(ms *ModuleIDs) UnmarshalYAML`;
* `TestModuleKindUnmarshalYAML`;
* `TestModuleIDsUnmarshalYAML`.
