+++ pkg/config/dict_test.go
@@ -20,7 +20,6 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/zclconf/go-cty-debug/ctydebug"
 	"github.com/zclconf/go-cty/cty"
-	"gopkg.in/yaml.v3"
 )
 
 func TestZeroValueValid(t *testing.T) {
@@ -78,109 +77,6 @@ func TestItemsAreCopy(t *testing.T) {
 	want := cty.StringVal("fuji")
 	got := d.Get("apple")
 	if diff := cmp.Diff(want, got, ctydebug.CmpOptions); diff != "" {
-		t.Errorf("diff (-want +got):\n%s", diff)
-	}
-}
-
-func TestYAMLDecode(t *testing.T) {
-	yml := `
-s1: "red"
-s2: pink
-m1: {}	
-m2:
-  m2f1: green
-  m2f2: [1, 0.2, -3, false]
-  gv: $(vars.gold)
-  mv: $(lime.bloom)
-  hl: ((3 + 9))
-`
-	want := Dict{}
-	want.
-		Set("s1", cty.StringVal("red")).
-		Set("s2", cty.StringVal("pink")).
-		Set("m1", cty.EmptyObjectVal).
-		Set("m2", cty.ObjectVal(map[string]cty.Value{
-			"m2f1": cty.StringVal("green"),
-			"m2f2": cty.TupleVal([]cty.Value{
-				cty.NumberIntVal(1),
-				cty.NumberFloatVal(0.2),
-				cty.NumberIntVal(-3),
-				cty.False,
-			}),
-			"gv": MustParseExpression("var.gold").AsValue(),
-			"mv": MustParseExpression("module.lime.bloom").AsValue(),
-			"hl": MustParseExpression("3 + 9").AsValue(),
-		}))
-	var got Dict
-	if err := yaml.Unmarshal([]byte(yml), &got); err != nil {
-		t.Fatalf("failed to decode: %v", err)
-	}
-	if diff := cmp.Diff(want.Items(), got.Items(), ctydebug.CmpOptions); diff != "" {
-		t.Errorf("diff (-want +got):\n%s", diff)
-	}
-}
-
-func TestMarshalYAML(t *testing.T) {
-	d := Dict{}
-	d.
-		Set("s1", cty.StringVal("red")).
-		Set("m1", cty.EmptyObjectVal).
-		Set("m2", cty.ObjectVal(map[string]cty.Value{
-			"m2f1": cty.StringVal("green"),
-			"m2f2": cty.TupleVal([]cty.Value{
-				cty.NumberIntVal(1),
-				cty.NumberFloatVal(0.2),
-				cty.NumberIntVal(-3),
-				cty.False,
-				MustParseExpression("7 + 4").AsValue(),
-			}),
-		}))
-	want := map[string]interface{}{
-		"s1": "red",
-		"m1": map[string]interface{}{},
-		"m2": map[string]interface{}{
-			"m2f1": "green",
-			"m2f2": []interface{}{1.0, 0.2, -3.0, false, "((7 + 4))"},
-		},
-	}
-	got, err := d.MarshalYAML()
-	if err != nil {
-		t.Fatalf("failed to marshal: %v", err)
-	}
-	if diff := cmp.Diff(want, got); diff != "" {
-		t.Errorf("diff (-want +got):\n%s", diff)
-	}
-}
-
-func TestYAMLMarshalIntAsInt(t *testing.T) {
-	d := Dict{}
-	d.Set("zebra", cty.NumberIntVal(5))
-	want := "zebra: 5\n"
-	got, err := yaml.Marshal(d)
-	if err != nil {
-		t.Fatalf("failed to marshal: %v", err)
-	}
-	if diff := cmp.Diff(want, string(got)); diff != "" {
-		t.Errorf("diff (-want +got):\n%s", diff)
-	}
-}
-
-func TestYAMLDecodeWithAlias(t *testing.T) {
-	yml := `
-pony: &passtime
-- eat
-- sleep
-zebra: *passtime
-`
-	want := Dict{}
-	want.
-		Set("pony", cty.TupleVal([]cty.Value{cty.StringVal("eat"), cty.StringVal("sleep")})).
-		Set("zebra", cty.TupleVal([]cty.Value{cty.StringVal("eat"), cty.StringVal("sleep")}))
-	var got Dict
-	if err := yaml.Unmarshal([]byte(yml), &got); err != nil {
-		t.Fatalf("failed to decode: %v", err)
-	}
-	if diff := cmp.Diff(want.Items(), got.Items(), ctydebug.CmpOptions); diff != "" {
 		t.Errorf("diff (-want +got):\n%s", diff)
 	}
 }
