+++ pkg/config/expand_test.go
@@ -211,7 +211,7 @@ func (s *MySuite) TestApplyUseModules(c *C) {
 		using := Module{
 			ID:     "usingModule",
 			Source: "path/using",
-			Use:    []ModuleID{"usedModule"},
+			Use:    ModuleIDs{"usedModule"},
 		}
 		used := Module{ID: "usedModule", Source: "path/used"}
