+++ pkg/config/yaml.go
@@ -17,10 +17,14 @@ package config
 import (
 	"bufio"
 	"bytes"
+	"encoding/json"
 	"fmt"
 	"os"
 	"regexp"
 
+	"github.com/zclconf/go-cty/cty"
+	"github.com/zclconf/go-cty/cty/gocty"
+	ctyJson "github.com/zclconf/go-cty/cty/json"
 	"gopkg.in/yaml.v3"
 )
 
@@ -156,4 +160,145 @@ type nodeCapturer struct{ n *yaml.Node }
 func (c *nodeCapturer) UnmarshalYAML(n *yaml.Node) error {
 	c.n = n
 	return nil
+}
+
+// UnmarshalYAML implements a custom unmarshaler from YAML string to ModuleKind
+func (mk *ModuleKind) UnmarshalYAML(n *yaml.Node) error {
+	var kind string
+	err := n.Decode(&kind)
+	if err == nil && IsValidModuleKind(kind) {
+		mk.kind = kind
+		return nil
+	}
+	return fmt.Errorf("line %d: kind must be \"packer\" or \"terraform\" or removed from YAML", n.Line)
+}
+
+// MarshalYAML implements a custom marshaler from ModuleKind to YAML string
+func (mk ModuleKind) MarshalYAML() (interface{}, error) {
+	return mk.String(), nil
+}
+
+// UnmarshalYAML is a custom unmarshaler for Module.Use, that will print nice error message.
+func (ms *ModuleIDs) UnmarshalYAML(n *yaml.Node) error {
+	var ids []ModuleID
+	if err := n.Decode(&ids); err != nil {
+		return fmt.Errorf("line %d: `use` must be a list of module ids", n.Line)
+	}
+	*ms = ids
+	return nil
+}
+
+// YamlValue is wrapper around cty.Value to handle YAML unmarshal.
+type YamlValue struct {
+	v cty.Value
+}
+
+// Unwrap returns wrapped cty.Value.
+func (y YamlValue) Unwrap() cty.Value {
+	return y.v
+}
+
+// UnmarshalYAML implements custom YAML unmarshaling.
+func (y *YamlValue) UnmarshalYAML(n *yaml.Node) error {
+	var err error
+	switch n.Kind {
+	case yaml.ScalarNode:
+		err = y.unmarshalScalar(n)
+	case yaml.MappingNode:
+		err = y.unmarshalObject(n)
+	case yaml.SequenceNode:
+		err = y.unmarshalTuple(n)
+	default:
+		err = fmt.Errorf("line %d: cannot decode node with unknown kind %d", n.Line, n.Kind)
+	}
+	return err
+}
+
+func (y *YamlValue) unmarshalScalar(n *yaml.Node) error {
+	var s interface{}
+	if err := n.Decode(&s); err != nil {
+		return err
+	}
+	ty, err := gocty.ImpliedType(s)
+	if err != nil {
+		return err
+	}
+	if y.v, err = gocty.ToCtyValue(s, ty); err != nil {
+		return err
+	}
+
+	if l, is := IsYamlExpressionLiteral(y.v); is { // HCL literal
+		var e Expression
+		if e, err = ParseExpression(l); err != nil {
+			return err
+		}
+		y.v = e.AsValue()
+	} else if y.v.Type() == cty.String && hasVariable(y.v.AsString()) { // "simple" variable
+		e, err := SimpleVarToExpression(y.v.AsString())
+		if err != nil {
+			return err
+		}
+		y.v = e.AsValue()
+	}
+	return nil
+}
+
+func (y *YamlValue) unmarshalObject(n *yaml.Node) error {
+	var my map[string]YamlValue
+	if err := n.Decode(&my); err != nil {
+		return err
+	}
+	mv := map[string]cty.Value{}
+	for k, y := range my {
+		mv[k] = y.v
+	}
+	y.v = cty.ObjectVal(mv)
+	return nil
+}
+
+func (y *YamlValue) unmarshalTuple(n *yaml.Node) error {
+	var ly []YamlValue
+	if err := n.Decode(&ly); err != nil {
+		return err
+	}
+	lv := []cty.Value{}
+	for _, y := range ly {
+		lv = append(lv, y.v)
+	}
+	y.v = cty.TupleVal(lv)
+	return nil
+}
+
+// UnmarshalYAML implements custom YAML unmarshaling.
+func (d *Dict) UnmarshalYAML(n *yaml.Node) error {
+	var m map[string]YamlValue
+	if err := n.Decode(&m); err != nil {
+		return err
+	}
+	for k, y := range m {
+		d.Set(k, y.v)
+	}
+	return nil
+}
+
+// MarshalYAML implements custom YAML marshaling.
+func (d Dict) MarshalYAML() (interface{}, error) {
+	o, _ := cty.Transform(d.AsObject(), func(p cty.Path, v cty.Value) (cty.Value, error) {
+		if e, is := IsExpressionValue(v); is {
+			return e.makeYamlExpressionValue(), nil
+		}
+		return v, nil
+	})
+
+	j := ctyJson.SimpleJSONValue{Value: o}
+	b, err := j.MarshalJSON()
+	if err != nil {
+		return nil, fmt.Errorf("failed to marshal JSON: %v", err)
+	}
+	var g interface{}
+	err = json.Unmarshal(b, &g)
+	if err != nil {
+		return nil, fmt.Errorf("failed to unmarshal JSON: %v", err)
+	}
+	return g, nil
 }
